insea-chatbot-main/Llm/requirements.txt
langchain
openai
transformers
opensearch-py
sentence-transformers
tiktoken
opensearch-py
youtube-transcript-api
google-generativeai
load_dotenv
